+++ src/gui-remote/collectionwidget.h
@@ -1,5 +1,5 @@
 /*
-    Copyright (C) 2016-2018, Kevin Andre <hyperquantum@gmail.com>
+    Copyright (C) 2016-2020, Kevin Andre <hyperquantum@gmail.com>
 
     This file is part of PMP (Party Music Player).
 
@@ -44,6 +44,7 @@ namespace PMP {
         void setConnection(ServerConnection* connection);
 
     private slots:
+        void highlightTracksIndexChanged(int index);
         void collectionContextMenuRequested(const QPoint& position);
 
     private: